==> 03_Iceberg_PySpark_Quickstart.ipynb <==
# Using Apache Iceberg with Spark 3 in CML

The official documentation for Apache Iceberg with Spark is located at [this link](https://iceberg.apache.org/#getting-started/#using-iceberg-in-spark-3)

For a full list of Apache Iceberg terms, please visit [this link](https://iceberg.apache.org/#terms/)

import cml.data_v1 as cmldata

# Sample in-code customization of spark configurations
#from pyspark import SparkContext
#SparkContext.setSystemProperty('spark.executor.cores', '1')
#SparkContext.setSystemProperty('spark.executor.memory', '2g')

CONNECTION_NAME = "go01-aw-dl"
conn = cmldata.get_connection(CONNECTION_NAME)
spark = conn.get_spark_session()

import os

username = os.environ["PROJECT_OWNER"]

#### You can use simple Spark SQL commands to create Spark tables as you always have. Just make sure to specify the USING iceberg clause.

spark.sql("CREATE TABLE IF NOT EXISTS {}_ice (id bigint, data string) USING iceberg".format(username))

#### To select a specific table snapshot or the snapshot at some time, Iceberg supports two Spark read options:

* snapshot-id selects a specific table snapshot
* as-of-timestamp selects the current snapshot at a timestamp, in milliseconds

#### You can view all snapshots associated with the table

spark.sql("SELECT * FROM {}_ice".format(username))

#### Or a full table version history 

spark.read.format("iceberg").load("spark_catalog.default.{}_ice.history".format(username)).show(20, False)

##### A manifest file is a metadata file that lists a subset of data files that make up a snapshot.

##### Each data file in a manifest is stored with a partition tuple, column-level stats, and summary information used to prune splits during scan planning.

##### To show a table’s file manifests and each file’s metadata, run:

spark.read.format("iceberg").load("spark_catalog.default.{}_ice.manifests".format(username)).show(5, False)

## Time Travel

### Using snapshots as shown above, we can insert some data into the table and roll back to its original state

# Insert using Iceberg format
spark.sql("INSERT INTO spark_catalog.default.{}_ice VALUES (1, 'x'), (2, 'y'), (3, 'z')".format(username))

# Query using select
spark.sql("SELECT * FROM spark_catalog.default.{}_ice".format(username)).show()

from datetime import datetime

# current date and time
now = datetime.now()

timestamp = datetime.timestamp(now)
print("timestamp =", timestamp)

# Query using a point in time
df = spark.read.option("as-of-timestamp", int(timestamp*1000)).format("iceberg").load("spark_catalog.default.{}_ice".format(username))
df.show(100)

# Insert using Iceberg format
spark.sql("INSERT INTO spark_catalog.default.{}_ice VALUES (1, 'd'), (2, 'e'), (3, 'f')".format(username))

#### Let's insert more data into the table

# Insert using Iceberg format
import string
import random

for i in range(25):
    number = random.randint(0, 10)
    letter = random.choice(string.ascii_letters)
    spark.sql("INSERT INTO spark_catalog.default.{}_ice VALUES ({}, '{}')".format(username, number, letter))

#### Now let's access the data again. Let's access it with the same timestemp as before. Notice we have a smaller number of rows than we just inserted.

# Query using a point in time
df = spark.read.option("as-of-timestamp", int(timestamp*1000)).format("iceberg").load("spark_catalog.default.{}_ice".format(username))
df.show(100)

### Observe that many new Snapshots have been created

spark.read.format("iceberg").load("spark_catalog.default.{}_ice.history".format(username)).show(10, False)

### You can also query the table in its previous state as of a specific partition.

#### Copy paste a partition_id from above and paste it in the next Spark command

spark.read\
    .option("snapshot-id", 3789197272971574187)\
    .table("spark_catalog.default.{}_ice".format(username)).show()

### The Iceberg API allows you to create tables from Spark Dataframes, and more

new_df = spark.sql("SELECT * FROM spark_catalog.default.{}_ice".format(username)).sample(fraction=0.5, seed=3)

new_df.dtypes

new_df.show(10)
